make require() paths more consistent in Utils.loadNamespace
This only has an impact if passing a subdirectory to the function
i.e. "ecs/systems" or whathaveyou
require() will store its result in package.loaded, but different IDs
will be stored under different keys in that table, even if they
point to the same file
i.e. "ecs/systems.collision" vs "ecs.systems.collision"
Because one would not have the same identity as the other,
require()ing the latter would return a different systemClass
than the former, therefore could not be used by world:getSystem

This change serves to make it easier to use require() to fetch
a given systemClass to find within a world, by making loadNamespace
convert the former ID to the latter